+++ SpehsEngine/Core/Murmur3.cpp
@@ -70,7 +70,7 @@ namespace se
 
 	//-----------------------------------------------------------------------------
 
-	uint32_t murmurHash3_x86_32(const void * key, int len, uint32_t seed)
+	uint32_t murmurHash3_x86_32(const void * key, size_t len, uint32_t seed)
 	{
 		const uint8_t * data = (const uint8_t*)key;
 		const int nblocks = len / 4;
@@ -126,7 +126,7 @@ namespace se
 
 	//-----------------------------------------------------------------------------
 
-	void murmurHash3_x86_128(const void * key, const int len, uint32_t seed, void * out)
+	void murmurHash3_x86_128(const void * key, const size_t len, uint32_t seed, void * out)
 	{
 		const uint8_t * data = (const uint8_t*)key;
 		const int nblocks = len / 16;
@@ -231,7 +231,7 @@ namespace se
 
 	//-----------------------------------------------------------------------------
 
-	void murmurHash3_x64_128(const void * key, const int len, const uint32_t seed, void * out)
+	void murmurHash3_x64_128(const void * key, const size_t len, const uint32_t seed, void * out)
 	{
 		const uint8_t * data = (const uint8_t*)key;
 		const int nblocks = len / 16;
